added fund transfer functionality
added fund transfer functionality

+++ src/main/java/com/scrotifybanking/scrotifybanking/repository/TransactionRepository.java
@@ -1,9 +1,12 @@
 package com.scrotifybanking.scrotifybanking.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import com.scrotifybanking.scrotifybanking.entity.Transaction;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
+/**
+ * The interface Transaction repository.
+ */
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+
 
 }
